# tests/contract/api-gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Multi-Agent PPM Platform",
    "description": "AI-native Project Portfolio Management platform with 25 specialized agents",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/vendors/{vendor_id}/research": {
      "post": {
        "tags": [
          "vendor-research"
        ],
        "summary": "Research Vendor",
        "description": "Trigger external vendor research using the Vendor & Procurement agent.",
        "operationId": "research_vendor_v1_vendors__vendor_id__research_post",
        "parameters": [
          {
            "name": "vendor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Vendor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorResearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VendorResearchRequest": {
        "properties": {
          "domain": {
            "type": "string",
            "minLength": 2,
            "title": "Domain",
            "description": "Procurement domain or category."
          },
          "vendor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Name",
            "description": "Optional vendor name override."
          }
        },
        "type": "object",
        "required": [
          "domain"
        ],
        "title": "VendorResearchRequest"
      },
      "VendorResearchResponse": {
        "properties": {
          "vendor_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Id"
          },
          "vendor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vendor Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "insights": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Insights"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources"
          },
          "used_external_research": {
            "type": "boolean",
            "title": "Used External Research"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "vendor_id",
          "vendor_name",
          "summary",
          "insights",
          "sources",
          "used_external_research"
        ],
        "title": "VendorResearchResponse"
      },
...
    }
  }
}